Add AUR package upd72020x-fw
The package provides the module xhci_pci. mkinitcpio displays a warning
due to the module being missing. It was installed when troubleshooting a
keyboard layout error in grub but it turned out to be unrelated. No
issue do to the missing module is known but might as well keep it since
the laptop has a xhci controller.